requirements.txt
matplotlib==3.4.3
numpy==1.21.6
pandas==1.3.5
Pillow==8.3.1
Pillow==10.3.0
torch==1.10.2
torchvision==0.11.3


# conda install unzip
# conda install tensorflow-gpu
# conda install pytorch torchvision pytorch-cuda=11.8 -c pytorch -c nvidia
# nvitop